Check the file name as well to ensure it delete the correct files.

+++ scripts/clean_old_logs.py
@@ -5,6 +5,6 @@ for examplePath, subFolders, files in os.walk(".") :
      now = datetime.datetime.now()
      dur = datetime.timedelta(days=31)
      history = now - dur
-     print os.path.getmtime(examplePath+"/"+f)
-     if history > datetime.datetime.fromtimestamp(os.path.getmtime(examplePath+"/"+f)) :
-             os.remove(examplePath+"/"+f)
+     if history > datetime.datetime.fromtimestamp(os.path.getmtime(examplePath+"/"+f)) \
+          and f.startswith("nightly_") and (not "history" in f) :
+               os.remove(examplePath+"/"+f)
